fortune50_ai_app/requirements.txt
# Core application dependencies
kafka-python==2.0.2
mlflow==2.8.1
hvac==1.2.1
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3

# Security and authentication
PyJWT==2.8.0
cryptography==41.0.7

# HTTP and API dependencies
requests==2.31.0
urllib3==2.0.7

# Configuration and environment
python-dotenv==1.0.0

# Monitoring and observability
prometheus-client==0.19.0

# Additional ML and data processing
joblib==1.3.2
scipy==1.11.4

# Development and testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
black==23.11.0

# Container and deployment
gunicorn==21.2.0